requirements.txt
Flask>=3.0.0,<4.0.0
PyJWT>=2.0.0,<3.0.0
gunicorn==20.1.0
pytest>=6.0
pytest-cov>=2.0
flake8==4.0.1
importlib_metadata<5.0